requirements.txt
anthropic
python-dotenv
sqlalchemy